feat: add base layout
